addTagToVideo
