+++ src/resolvers/Mutation/addTagToVideo/resolver.ts
@@ -1,67 +1,14 @@
-import { Tag, UserRole, Video, VideoTag, VideoTagEventType } from "@prisma/client";
-import { ulid } from "ulid";
+import { UserRole } from "@prisma/client";
 
-import { Result } from "../../../utils/Result.js";
 import { AddTagToVideoFailedMessage, MutationResolvers } from "../../graphql.js";
 import { parseGqlID2 } from "../../id.js";
 import { ResolverDeps } from "../../index.js";
 import { TagModel } from "../../Tag/model.js";
 import { VideoModel } from "../../Video/model.js";
 import { addTagToVideoInNeo4j } from "./neo4j.js";
+import { add } from "./prisma.js";
 
-export const add = async (
-  prisma: ResolverDeps["prisma"],
-  { authUserId: userId, videoId, tagId }: { authUserId: string; videoId: string; tagId: string }
-): Promise<Result<"EXISTS_TAGGING", VideoTag & { video: Video; tag: Tag }>> => {
-  const exists = await prisma.videoTag.findUnique({ where: { videoId_tagId: { tagId, videoId } } });
-  if (exists && !exists.isRemoved) return { status: "error", error: "EXISTS_TAGGING" };
-
-  if (exists) {
-    // reattach
-    const tagging = await prisma.videoTag.update({
-      where: { id: exists.id },
-      data: {
-        isRemoved: false,
-        events: {
-          create: {
-            userId,
-            type: VideoTagEventType.REATTACH,
-            payload: {},
-          },
-        },
-      },
-      include: { video: true, tag: true },
-    });
-
-    return {
-      status: "ok",
-      data: tagging,
-    };
-  } else {
-    // attach
-    const id = ulid();
-    const tagging = await prisma.videoTag.create({
-      data: {
-        id,
-        tagId,
-        videoId,
-        isRemoved: false,
-        events: {
-          create: {
-            userId,
-            type: VideoTagEventType.ATTACH,
-            payload: {},
-          },
-        },
-      },
-      include: { video: true, tag: true },
-    });
-
-    return { status: "ok", data: tagging };
-  }
-};
-
-export const addTagToVideo = ({ neo4j, prisma, logger }: Pick<ResolverDeps, "prisma" | "neo4j" | "logger">) =>
+export const resolverAddTagToVideo = ({ neo4j, prisma, logger }: Pick<ResolverDeps, "prisma" | "neo4j" | "logger">) =>
   (async (_parent, { input: { tagId: tagGqlId, videoId: videoGqlId } }, { user }, info) => {
     if (!user || (user?.role !== UserRole.EDITOR && user?.role !== UserRole.ADMINISTRATOR))
       return {
